// crates/desktop/src/generated/dto/OperationBatchResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { OperationResultDto } from "./OperationResultDto";

export interface OperationBatchResponse {
	success_count: number;
	failed_count: number;
	results: Array<OperationResultDto>;
}