// crates/desktop/src/generated/dto/TransportDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type TransportDto = { "type": "stdio", command: string, args: Array<string>, env: { [key in string]: string } | null, timeout: number | null, } | { "type": "sse", url: string, headers: { [key in string]: string } | null, timeout: number | null, } | { "type": "streamable_http", url: string, headers: { [key in string]: string } | null, timeout: number | null, };